refactor: simplify action-menu
