+++ packages/components/src/components/action-bar/utils.ts
@@ -52,7 +52,3 @@ export const overflowActions = ({
     group.manager.component.requestUpdate();
   });
 };
-
-export function isAction(el: Element | null): el is Action["el"] {
-  return el?.tagName === "CALCITE-ACTION";
-}
