+++ packages/components/src/components/action/action.tsx
@@ -1,15 +1,6 @@
 // @ts-strict-ignore
 import { createRef } from "lit/directives/ref.js";
-import {
-  LitElement,
-  property,
-  h,
-  method,
-  JsxNode,
-  Fragment,
-  LuminaJsx,
-  createEvent,
-} from "@arcgis/lumina";
+import { LitElement, property, h, method, JsxNode, Fragment, LuminaJsx } from "@arcgis/lumina";
 import { guid } from "../../utils/guid";
 import { createObserver } from "../../utils/observers";
 import { getIconScale } from "../../utils/component";
@@ -210,16 +201,6 @@ export class Action extends LitElement implements FormOwner {
 
   //#endregion
 
-  //#region Events
-
-  /**
-   * Fires when the action's button is being pressed down.
-   * @internal
-   */
-  calciteInternalActionMouseDown = createEvent({ cancelable: false });
-
-  //#endregion
-
   //#region Lifecycle
 
   override connectedCallback(): void {
@@ -243,10 +224,6 @@ export class Action extends LitElement implements FormOwner {
     } else if (type === "reset") {
       resetForm(this);
     }
-  }
-
-  private handleMouseDown(): void {
-    this.calciteInternalActionMouseDown.emit();
   }
 
   //#endregion
@@ -400,7 +377,6 @@ export class Action extends LitElement implements FormOwner {
         disabled={disabled}
         id={buttonId}
         onClick={this.handleClick}
-        onMouseDown={this.handleMouseDown}
         ref={this.buttonRef}
         role={this.aria?.role}
       >
